requirements.txt
flask
flask-cors
pymongo
dnspython
simplejson
python-dotenv